backend/requirements.txt
boto3
awscli
Pillow
Flask
flask-cors